Document APIClient init

+++ cads_api_client/catalogue.py
@@ -85,7 +85,7 @@ class Catalogue:
     retry_options: dict[str, Any]
     request_options: dict[str, Any]
     download_options: dict[str, Any]
-    sleep_max: int
+    sleep_max: float
     cleanup: bool
     force_exact_url: bool = False
